requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_community
streamlit
python-telegram-bot